# openapi/railwaysMySQL-api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8081
  description: Generated server url
paths:
...
  /api/routes/{routeId}:
    get:
      tags:
      - routes
      summary: Get an existing route
      operationId: getSingleRoute
      parameters:
      - name: routeId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "404":
          description: No route to get
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Route'
    put:
      tags:
      - routes
      summary: Update an existing route
      operationId: updateSingleRoute
      parameters:
      - name: routeId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Route'
        required: true
      responses:
        "404":
          description: No route to update
        "200":
          description: Route successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Route'
    delete:
      tags:
      - routes
      summary: Delete an existing route
      operationId: deleteRoute
      parameters:
      - name: routeId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "404":
          description: No route to delete
        "204":
          description: No content
...
  /api/trains/{trainId}/routes/{routeId}:
    post:
      tags:
      - routes-trains
      summary: Add a new association between route and train
      operationId: createTrainRouteAssociation
      parameters:
      - name: trainId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      - name: routeId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "409":
          description: Already exists an association between ids in path
        "201":
          description: Created
    delete:
      tags:
      - routes-trains
      summary: Delete an existing association between train and route
      operationId: deleteTrainRouteAssociation
      parameters:
      - name: trainId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      - name: routeId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "404":
          description: No association to delete
        "204":
          description: No content
  /api/routes:
    get:
      tags:
      - routes
      summary: Get routes list with optional filters
      operationId: findRoutePage
      parameters:
      - name: destinationStationName
        in: query
        required: false
        schema:
          type: string
      - name: startStationName
        in: query
        required: false
        schema:
          type: string
      - name: binaryType
        in: query
        required: false
        schema:
          type: string
      - name: minimumLength
        in: query
        required: false
        schema:
          type: number
          format: double
      - name: maxLength
        in: query
        required: false
        schema:
          type: number
          format: double
      - name: offset
        in: query
        required: true
        schema:
          type: integer
          format: int32
      - name: limit
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page'
        "400":
          description: Bad request
    post:
      tags:
      - routes
      summary: Add a new route
      operationId: createRoute
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Route'
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Route'
        "409":
          description: Already exists a route with id in body
...
components:
  schemas:
...
    Route:
      type: object
      properties:
        id:
          type: integer
          format: int32
        binaryType:
          type: string
        startStation:
          type: string
        destinationStation:
          type: string
        length:
          type: number
          format: double
        trains:
          uniqueItems: true
          type: array
          items:
            properties:
              trainIDs:
                type: number
                format: int32
...
    Page:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            type: object
        number:
          type: integer
          format: int32
        sort:
          $ref: '#/components/schemas/Sort'
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: '#/components/schemas/Pageable'
        empty:
          type: boolean
    Pageable:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          $ref: '#/components/schemas/Sort'
        pageSize:
          type: integer
          format: int32
        pageNumber:
          type: integer
          format: int32
        unpaged:
          type: boolean
        paged:
          type: boolean
    Sort:
      type: object
      properties:
        sorted:
          type: boolean
        unsorted:
          type: boolean
        empty:
          type: boolean